# server/docs/swagger.yaml
basePath: /
definitions:
  entity.Actor:
    properties:
      birthday:
        example: "1990-01-01"
        type: string
      name:
        example: John Doe
        type: string
      sex:
        example: male
        type: string
    type: object
  entity.ActorFilms:
    properties:
      birthday:
        type: string
      films:
        items:
          $ref: '#/definitions/entity.Film'
        type: array
      name:
        type: string
      sex:
        type: string
    type: object
  entity.Film:
    properties:
      description:
        example: A mind-bending thriller
        type: string
      name:
        example: Inception
        type: string
      rating:
        example: 8
        maximum: 10
        minimum: 0
        type: integer
      releaseDay:
        example: "2010-07-16"
        type: string
    type: object
  entity.UpdateActorInput:
    properties:
      birthday:
        example: "1992-12-12"
        type: string
      name:
        example: John Doe
        type: string
      sex:
        example: female
        type: string
    type: object
  entity.UpdateFilmInput:
    properties:
      description:
        example: New description
        type: string
      name:
        example: Inception
        type: string
      rating:
        example: 6
        maximum: 10
        minimum: 0
        type: integer
      releaseDay:
        example: "2010-07-16"
        type: string
    type: object
  entity.User:
    properties:
      password:
        example: strongPassword
        type: string
      role:
        example: admin
        type: string
      username:
        example: john_doe
        type: string
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  title: Film library
  version: "1.0"
paths:
  /api/actor:
    post:
      consumes:
      - application/json
      description: Creates a new actor.
      parameters:
      - description: 'Data of the new actor (example: {''name'': ''John Doe'', ''sex'':
          ''male'', ''birthday'': ''1999-10-12''})'
        in: body
        name: actor
        required: true
        schema:
          $ref: '#/definitions/entity.Actor'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created actor
          schema:
            type: integer
        "400":
          description: Invalid request data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create actor
      tags:
      - Actors
  /api/actor/{id}:
    delete:
      description: Deletes an actor with the specified ID.
      parameters:
      - description: Actor ID to delete
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Actor deleted successfully
          schema:
            type: string
        "400":
          description: Invalid actor ID param
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete actor by ID
      tags:
      - Actors
    patch:
      consumes:
      - application/json
      description: Updates an actor with the specified ID based on the data passed
        in the request body.
      parameters:
      - description: Actor ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Data of the actor to update
        in: body
        name: actor
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateActorInput'
      produces:
      - application/json
      responses:
        "200":
          description: Actor updated successfully
          schema:
            type: string
        "400":
          description: Invalid actor ID param or request data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update actor by ID
      tags:
      - Actors
  /api/actors:
    get:
      consumes:
      - application/json
      description: Returns actors with their associated films based on the provided
        actor IDs.
      parameters:
      - collectionFormat: csv
        description: Actor IDs
        in: query
        items:
          type: integer
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/entity.ActorFilms'
              type: array
            type: array
        "400":
          description: Invalid request or invalid actor ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get actors with associated films
      tags:
      - Actors
  /api/film:
    post:
      consumes:
      - application/json
      description: Creates a new film.
      parameters:
      - description: Data of the new film
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/entity.Film'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created film
          schema:
            type: integer
        "400":
          description: Invalid request data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create film
      tags:
      - Films
  /api/film/{id}:
    delete:
      description: Deletes a film with the specified ID.
      parameters:
      - description: Film ID to delete
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Film deleted successfully
          schema:
            type: string
        "400":
          description: Invalid film ID param
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete film by ID
      tags:
      - Films
    patch:
      consumes:
      - application/json
      description: Updates a film with the specified ID based on the data passed in
        the request body.
      parameters:
      - description: Film ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Data of the film to update
        in: body
        name: film
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateFilmInput'
      produces:
      - application/json
      responses:
        "200":
          description: Film updated successfully
          schema:
            type: string
        "400":
          description: Invalid film ID param or request data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update film by ID
      tags:
      - Films
  /api/films/fragments:
    get:
      consumes:
      - application/json
      description: This endpoint allows searching for films based on actor name and
        film name fragments.
      parameters:
      - description: Film name fragment
        in: query
        name: filmNameFr
        type: string
      - description: Actor name fragment
        in: query
        name: actorNameFr
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/entity.Film'
              type: array
            type: array
        "400":
          description: Invalid request or invalid actor ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Search for films by actor name and film name fragments
      tags:
      - Films
  /api/films/sorted:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves films sorted by the specified criteria
        and associated with the provided actor IDs.
      parameters:
      - description: 'Sort type: rating, date, name.'
        in: query
        name: sortType
        type: string
      - collectionFormat: csv
        description: Actor IDs
        in: query
        items:
          type: integer
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/entity.Film'
              type: array
            type: array
        "400":
          description: Invalid request or invalid actor ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve films sorted by criteria and associated with actors
      tags:
      - Films
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Signs in an existing user.
      parameters:
      - description: User credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            type: string
        "400":
          description: Invalid request data
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Sign in a user
      tags:
      - Authentication
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Creates a new user account.
      parameters:
      - description: 'User data example: ''{''username'': ''example_user'', ''password'':
          ''example_password'', ''role'': ''Enums(admin, user)''}'
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: ID of the created user
          schema:
            type: string
        "400":
          description: Invalid request data
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Sign up a new user
      tags:
      - Authentication
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT token (example: Bearer <your_jwt_token_here>)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"